npm: adding more javascript to the pile (retry)
So sad that this language is so BAD, and poorly made, that it can't even be bothered to allow capturing references for closures. Useless. Why would it want to stop me anyway? The language is single-threaded, so there is even less risk of a data race, it's just another stupid and incompetent decision. And yet another example of morons deciding what's best, and getting in the way of other developers. At least provide a proper alternative if you're going to stop me from doing instead of just telling me "you can't".