tests/testing-utilities: move expectToContainAllKeys to utilities
Thought I could keep this in a test file by it's self, but unfortunately it just causes test runtimes to bloat. So in the effort of not having to wait around any longer than I have to for tests to finish, I've gone ahead and done the hard work of moving it to where it should be. Even though it should be able to be messy, in a file that it doesn't need to be in without much fuss, but it just causes tests to re-run even if it's not their turn, because we don't live in the age of fucking sensible software unfortunately.