server: disable request speed limiter temporarily
This might be, single-handedly, the biggest factor on why tests were not only so flaky, but taking so long. After a single client makes a certain amount of requests, subsequent requests are slowed down significantly to prevent overloading the server. However, this has the issue of doing it indiscriminately, and without any way to disable with configuration, you end up having tests taking forever because of this.